Gg quannvph36493

+++ database/migrations/2024_08_24_142847_create_expenses_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('expenses', function (Blueprint $table) {
             $table->id();
-            $table->decimal('amount', 8, 2)->nullable();
+            $table->double('amount')->nullable();
             $table->string('title')->nullable();
             $table->text('description');
             $table->date('date')->nullable();
