+++ database/migrations/2024_08_24_142948_create_teachers_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->string('address', 255)->nullable();
             $table->date('birth_day')->nullable();
             $table->string('qualification')->nullable();
-            $table->decimal('hourly_rate', 8,2)->nullable();
+            $table->double('hourly_rate')->nullable();
             $table->date('enrollment_date')->default(Date::now());
             $table->timestamps();
             $table->softDeletes();
