+++ database/migrations/2025_09_24_211247_create_incomes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('incomes', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(Classes::class)->constrained();
-            $table->decimal('amount')->nullable();
+            $table->double('amount')->nullable();
             $table->string('title')->nullable();
             $table->text('description')->nullable();
             $table->date('date')->default(Date::now());
